comments entity

+++ src/modules/post/post.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { CreatePostDto } from 'src/modules/post/dto/create-post.dto';
 import { UpdatePostDto } from 'src/modules/post/dto/update-post.dto';
@@ -22,24 +22,30 @@ export class PostService {
     });
   }
 
-  async create(createPostDto: CreatePostDto) {
-    const post = this.postRepository.create(createPostDto);
+  async create(createPostDto: CreatePostDto, userId: number) {
+    const post = this.postRepository.create({ ...createPostDto, userId });
     return this.postRepository.save(post);
   }
 
-  async update(id: number, data: UpdatePostDto) {
+  async update(id: number, data: UpdatePostDto, userId: number) {
     const post = await this.findOne(id);
     if (!post) {
       throw new NotFoundException('Post not found');
+    }
+    if (userId !== post.userId) {
+      throw new ForbiddenException('You do not have permission to edit');
     }
     await this.postRepository.update(id, data);
     return this.findOne(id);
   }
 
-  async delete(id: number) {
+  async delete(id: number, userId: number) {
     const post = await this.findOne(id);
     if (!post) {
       throw new NotFoundException('Post not found');
+    }
+    if (userId !== post.userId) {
+      throw new ForbiddenException('You do not have permission to delete');
     }
     await this.postRepository.delete(id);
     return { message: 'Post deleted successfully' };
